Fixed problems w/ JPEG lossless not decompressing color images properly.  The previous checkin did not take into account not setting a colorspace conversion when lossless jpeg was being decompressed.  Also, cleaned up the extension points for the server.  They were not setup properly, so the codecs were not working properly for jpeg extended and lossless.  Also, cleaned up bit depth checks to determine which IJG decompressor to use.  There were some fixed settings for these, and changed these. Seems like we still might need some work in this area, because I didn't completely match this up to the way Offis was doing it.

+++ Dicom/Codec/Jpeg/DicomJpegCodec.cpp
@@ -89,8 +89,8 @@ void DicomJpegCodec::Encode(DicomUncompressedPixelData^ oldPixelData, DicomCompr
 	if (codec->Mode != JpegMode::Lossless) {
 		newPixelData->LossyImageCompressionMethod = "ISO_10918_1";
 		
-		double oldSize = oldPixelData->UncompressedFrameSize;
-		double newSize = newPixelData->GetCompressedFrameSize(0);
+		float oldSize = oldPixelData->UncompressedFrameSize;
+		float newSize = newPixelData->GetCompressedFrameSize(0);
 		String^ ratio = String::Format("{0:0.000}", oldSize / newSize);
 		newPixelData->LossyImageCompressionRatio = (float) oldSize / newSize;
 	}
